requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
APScheduler>=3.10.4
python-dotenv>=1.0.0
pandas>=2.1.0
rich>=13.7.0
tenacity>=8.2.3
groq>=0.4.0
